refactor: narrow import paths for types
