+++ pages/assignments/[assignmentId]/configs/[assignmentConfigId]/yaml.tsx
@@ -17,7 +17,7 @@ import { SlideOver } from "../../../../../components/SlideOver";
 import { useState } from "react";
 import AssignmentSupportingFilesUploader from "../../../../../components/AssignmentSupportingFilesUploader";
 import { FilesProvider } from "../../../../../contexts/assignmentSupportingFiles";
-import { Assignment, AssignmentConfig } from "@/types";
+import { Assignment, AssignmentConfig } from "@/types/tables";
 
 function SlideOverContent() {
   const { configSlideOver } = useLayoutState();
