+++ store/GuiBuilder/__tests__/pipelineEditorModel.test.ts
@@ -1,5 +1,5 @@
 import supportedStages from "@/constants/GuiBuilder/supportedStages";
-import { StageDataMap, StageKind, StageNode } from "@/types";
+import { StageDataMap, StageKind, StageNode } from "@/types/GuiBuilder";
 import { computed, createStore, thunkOn } from "easy-peasy";
 import "jest-extended";
 import * as uuid from "uuid";
