+++ store/GuiBuilder/model/pipelineEditorModel.ts
@@ -1,5 +1,5 @@
 import supportedStages, { SupportedStage } from "@/constants/GuiBuilder/supportedStages";
-import { Stage, StageNode } from "@/types";
+import { Stage, StageNode } from "@/types/GuiBuilder";
 import { deleteStageFromDeps } from "@/utils/GuiBuilder";
 import { coordQuad, dagConnect, sugiyama } from "d3-dag";
 import { action, Action, computed, Computed, thunk, Thunk, thunkOn, ThunkOn } from "easy-peasy";
