+++ utils/GuiBuilder/__tests__/config.test.ts
@@ -1,5 +1,5 @@
 import { defaultSettings } from "@/constants/GuiBuilder/defaults";
-import { Config, StageDataMap, StageDependencyMap, StageKind } from "@/types";
+import { Config, StageDataMap, StageDependencyMap, StageKind } from "@/types/GuiBuilder";
 import cloneDeep from "lodash/cloneDeep";
 import { configToYaml, parseConfigYaml } from "../config";
 import * as settingsUtils from "../settings";
