+++ utils/GuiBuilder/settings.ts
@@ -3,7 +3,7 @@
  */
 
 import { defaultSettings } from "@/constants/GuiBuilder/defaults";
-import type { Settings, SettingsLang, SettingsRaw } from "@/types";
+import { Settings, SettingsLang, SettingsRaw } from "@/types/GuiBuilder";
 import cloneDeep from "lodash/cloneDeep";
 import isEqual from "lodash/isEqual";
 import { v4 as uuidv4 } from "uuid";
